+++ vue/components/ui/molecules/button-upload/button-upload.vue
@@ -18,7 +18,7 @@
         <slot v-bind:open-modal="openModal" />
         <button-color
             class="button-upload"
-            v-bind:text="buttonText"
+            v-bind:text="buttonText || buttonTextComputed"
             v-bind:icon="'cloud-upload'"
             v-bind:alignment="'center'"
             v-bind:disabled="disabled"
@@ -57,15 +57,11 @@ export const ButtonUpload = {
     props: {
         buttonText: {
             type: String,
-            default: "Upload Files"
+            default: null
         },
         disabled: {
             type: Boolean,
             default: false
-        },
-        multiple: {
-            type: Boolean,
-            default: true
         },
         accept: {
             type: String,
@@ -86,6 +82,10 @@ export const ButtonUpload = {
                 disabled: this.draggingDisabled
             };
             return base;
+        },
+        buttonTextComputed() {
+            if (this.multiple) return "Upload Files";
+            return "Upload File";
         }
     },
     watch: {